[FE - SC]: Move the ethers.js based SC file into the /ethers-js directory
